Dr. Shore's Modified Bronze Challenge Complete:
Bronze Challenge:Renaming the Delete ButtonWhen deleting a row, a confirmation button appears labeled Delete. Change the label of this button to Remove.Cell ColorsUpdate the ItemCell to display the valueInDollars in green if the value is less than 50 and red if the value is greater than or equal to 50.